fix(v2): use json.Decoder with UseNumber() for proper integer handling
When unmarshalling SearchResultImpl, JSON numbers were being decoded as
float64, causing GetInt() to fail for integer metadata values. Using
json.Decoder with UseNumber() preserves number types as json.Number,
allowing NewDocumentMetadataFromMap to correctly distinguish between
integers and floats.

Also handle json.Number type in embeddings and scores parsing.

+++ pkg/api/v2/search_test.go
@@ -611,9 +611,9 @@ func TestSearchResultUnmarshal(t *testing.T) {
 		require.True(t, ok)
 		require.Equal(t, "AI", cat)
 
-		year, ok := result.Metadatas[0][0].GetFloat("year")
+		year, ok := result.Metadatas[0][0].GetInt("year")
 		require.True(t, ok)
-		require.Equal(t, float64(2023), year)
+		require.Equal(t, int64(2023), year)
 	})
 
 	t.Run("unmarshal empty result", func(t *testing.T) {
